refactor: tensorflow utils
