+++ queens/models/surrogates/bayesian_neural_network.py
@@ -21,8 +21,8 @@ import numpy as np
 import tensorflow_probability as tfp
 
 from queens.models.surrogates.surrogate import Surrogate
+from queens.utils.configure_tensorflow import configure_keras, configure_tensorflow
 from queens.utils.logger_settings import log_init_args
-from queens.utils.tensorflow import configure_keras, configure_tensorflow
 
 _logger = logging.getLogger(__name__)
