+++ queens/models/surrogates/gaussian_neural_network.py
@@ -21,9 +21,9 @@ import numpy as np
 import tensorflow_probability as tfp
 
 from queens.models.surrogates.surrogate import Surrogate
+from queens.utils.configure_tensorflow import configure_keras, configure_tensorflow
 from queens.utils.logger_settings import log_init_args
 from queens.utils.random_process_scaler import VALID_SCALER
-from queens.utils.tensorflow import configure_keras, configure_tensorflow
 from queens.utils.valid_options import get_option
 from queens.visualization.gaussian_neural_network_vis import plot_loss
