+++ queens/models/surrogates/heteroskedastic_gaussian_process.py
@@ -22,8 +22,8 @@ import tensorflow_probability as tfp
 from sklearn.cluster import KMeans
 
 from queens.models.surrogates.surrogate import Surrogate
+from queens.utils.configure_tensorflow import configure_keras, configure_tensorflow
 from queens.utils.logger_settings import log_init_args
-from queens.utils.tensorflow import configure_keras, configure_tensorflow
 
 _logger = logging.getLogger(__name__)
