+++ queens/models/surrogates/variational_gaussian_process.py
@@ -36,9 +36,9 @@ import numpy as np
 import tensorflow_probability as tfp
 
 from queens.models.surrogates.surrogate import Surrogate
+from queens.utils.configure_tensorflow import configure_keras, configure_tensorflow
 from queens.utils.gpflow import extract_block_diag, init_scaler, set_transform_function
 from queens.utils.logger_settings import log_init_args
-from queens.utils.tensorflow import configure_keras, configure_tensorflow
 
 _logger = logging.getLogger(__name__)
