+++ queens/utils/configure_tensorflow.py
